+++ processor/src/main/java/arez/processor/ArezProcessor.java
@@ -1097,6 +1097,11 @@ public final class ArezProcessor
     final boolean reportParameters = AnnotationsUtil.getAnnotationValueValue( annotation, "reportParameters" );
     final boolean reportResult = AnnotationsUtil.getAnnotationValueValue( annotation, "reportResult" );
     final boolean verifyRequired = AnnotationsUtil.getAnnotationValueValue( annotation, "verifyRequired" );
+    final boolean skipIfDisposed = AnnotationsUtil.getAnnotationValueValue( annotation, "skipIfDisposed" );
+    if ( skipIfDisposed && TypeKind.VOID != methodType.getReturnType().getKind() )
+    {
+      throw new ProcessorException( "@Action target must not return a value when skipIfDisposed=true", method );
+    }
     final ActionDescriptor action =
       new ActionDescriptor( component,
                             name,
@@ -1105,6 +1110,7 @@ public final class ArezProcessor
                             verifyRequired,
                             reportParameters,
                             reportResult,
+                            skipIfDisposed,
                             method,
                             methodType );
     component.getActions().put( action.getName(), action );